Refactor throwException() method. Replace codes with wrappers
